Tweaked Bullet to AlienShip collision detection and cleaned up code
Tweaked AlienShip to draw as a circle rather than ellipse for collision detection purposes.  AlienShip now only enters every 20 seconds if not active before hand. Added overloaded function of collisionDetection for bullets. Put the randomization of speed for asteroids in constructor.  Added randomStart function to AlienShip. Used Ben's code for AlienShip die function. Implemented the above to allow player bullets to destroy alienships, have an explosion at the area and reset the AlienShip.

+++ a3/explosion.pde
@@ -1,12 +1,15 @@
+// Created by CS on 29/03/20
+// Function to animate and draw the explosions
+// Only parameters needed are the origin x and y location of the explosion
 class explosion {
   
   int explosionOpacity;
   PVector explosionLoc[];
   int currentCycle = 0;
   int explosionLimit = 11;
-  int originX,originY;
+  float originX,originY;
   
-  explosion(int a, int b) {
+  explosion(float a, float b) {
     originX = a;
     originY = b;
